+++ go.mod
@@ -3,16 +3,16 @@ module github.com/hspak/gopm3
 go 1.18
 
 require (
-	github.com/gdamore/tcell/v2 v2.5.3
-	github.com/rivo/tview v0.0.0-20220916081518-2e69b7385a37
+	github.com/gdamore/tcell/v2 v2.5.4
+	github.com/rivo/tview v0.0.0-20230208211350-7dfff1ce7854
 )
 
 require (
 	github.com/gdamore/encoding v1.0.0 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
 	github.com/mattn/go-runewidth v0.0.14 // indirect
-	github.com/rivo/uniseg v0.4.2 // indirect
-	golang.org/x/sys v0.1.0 // indirect
-	golang.org/x/term v0.0.0-20221017184919-83659145692c // indirect
-	golang.org/x/text v0.4.0 // indirect
+	github.com/rivo/uniseg v0.4.3 // indirect
+	golang.org/x/sys v0.5.0 // indirect
+	golang.org/x/term v0.5.0 // indirect
+	golang.org/x/text v0.7.0 // indirect
 )
